LDA, HDA, and LSI topic models.
Hello, I was curious to see what topic models might put out for the interviews.  This script uses Latent Dirichlet Allocation, Hierarchical Dirichlet Allocation, and Latent Semantic Indexing.  The document is a result of their outputs.  Sorry about the obnoxious section titles on it, I was not concerned with appearance for it.